cleanup format strings for cfs_* messages
This does not change semantics on our current target platform
(x86_64) but is needed for porting it to other platforms.

The GCC on ARM, for example, complains about them.

For all:
* size_t use "%z*"
* off_t use "%j*"
* uint64_t use "PRI*64"

where * may be one of (X,d,u).

Also cast guint64 to uint64_t to allow use of a general, portable
format which also supports hex output as the GUINT64_FORMAT allows
decimal output only.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Cc: mir@datanom.net