buildsys: write control file into build directory
Else the first make {deb,dinstall} from a clean repo fails as we
generated the debian control file to the source debian/ folder.
Just write it directly to the build/debian directory, so we do not
clutter the source directory and build always with the up to date
control file.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>