pvecm create: remove rrp_mode parameter
I detected a bug where we overwrote the whole $interfaces variable
(and so all interfaces from the corosync config) if the 'rrp_mode'
param was set.

While this would be easy to with by changing the line to
$interfaces .= ...
I removed the whole rrp_mode parameter instead.

As:
a) I've seen no one running into this bug, so this parameter was not
   really used either way.
b) only the 'passive' is supported and works, 'active' has a whole
   lot of problems. If someone really wants it he should edit the
   corosync config file to achieve this

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>