pve-cluster service: remove $DAEMON_OPTS and environment file
"We use it to pass $DAEMON_OPTS variable to pmxcfs, which is empty per
default.

pmxcfs accepts following options at the moment:

> Usage:
>   pmxcfs [OPTION...]
>
> Help Options:
>   -h, --help           Show help options
>
> Application Options:
>   -d, --debug          Turn on debug messages
>   -f, --foreground     Do not daemonize server
>   -l, --local          Force local mode (ignore corosync.conf,
>   force quorum)

"help" can be safely ignored, as can "foreground" - it would break
the service as the Type is forking and thus systemd would expect that
the starting process returns rather quickly and kill it after the
timeout thinking the start failed when this is set.

Then there is "debug", while this is quite safe to use I do not
expect to find it in the wild. 1) it *spams* the logs in such a heavy
manner that it's useless for a user 2) if wished it can be
enabled/disable on the fly by writing 1/0 to /etc/pve/.debug (which
is what we tell people in docs, et al.  anyway) So this parameter is
also quite safe to ignore, IMO.

Then there is "local", now this can be enabled but is such evil
and wrong so that anybody setting it *permanently*, deserves to
be saved by not allowing him to do so. If somebody uses this we
should hope that he did out of not knowing better and is actually
thankful when he learns what this meant."
https://pve.proxmox.com/pipermail/pve-devel/2017-November/029527.html

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>