buildsys: autogen debug package and cleanup unecessary rules
don't do manually what the deb helpers do automatically and better.

Autogenerate the debug package, it includes now only the debugsymbols
without effectively duplicating all executables and libraries.
In the same step add a install file which installs our sysctl
settings, this is done together as it allows to skip some
intermediate steps, also the change is not to big, so it should be
possible to see whats going on.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>