pmxcfs: do not wait artificially when stopping
Most of PVE services profit from an up and running pmxcfs, thus
artificially prolonging its graceful termination is contra
productive.

As everything gets de-initialized nicely and gracefully this should
not be problematic, the restart is not the fastest anyhow anyway.

As this specific line has no git tracking information available
(imported from svn) I could not find if and what the reason for this
was.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>