use constants for IPC request types
Add a simple header with the constants as defines.
Use a simple awk script to translate this to an perl module with the
constants exported. awk is far easier to understand and maintain than
h2ph or h2xs, also their result is quite a mess for such a trivial
thing, IMO.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>