node join: use new corosync link parameters
Similar to the change to cluster creation use now also the
corosync-link definition for the rest of the cluster join/add calls.

As link0, former ring0, is not special anymore allow that it's not
passed and only default back to nodename if it's configured in the
totem section of the configuration.

As the 'join' and 'addnode' api paths are quite connected, do all in
one patch.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>