d/control: bump debhelper compat to >= 12
dh_systemd was enabled by default since level 10, with level 12 the
compat plugin does not exists anymore so enabling it manually results
in an error.

The dh_strip override is now obsolete too, as users need to go
through 5.4 AND 6.4 anyway on upgrade, and new installations do not
matter here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>